projet_fastAPI/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
redis
pydantic[email]
pydantic>=2
pydantic-settings>=2.2




